+++ test/interpolation_test.dart
@@ -49,6 +49,15 @@ void main() {
     expect(results, "SELECT * FROM t WHERE id=2 WHERE blob='{\"key\":\"value\"}'::jsonb AND blob='{\"a\":1}'::jsonb");
   });
 
+  test("Can both provide type and typecast", () {
+    var results = PostgreSQLFormat.substitute("SELECT * FROM t WHERE id=@id:int2::int4", {
+      "id": 2,
+      "blob": "{\"key\":\"value\"}"
+    });
+
+    expect(results, "SELECT * FROM t WHERE id=2::int4");
+  });
+
   test("String identifiers get escaped", () {
     var result = PostgreSQLFormat
         .substitute("@id:text @foo", {"id": "1';select", "foo": "3\\4"});